Refactor Interop to use NativeLibrary
