+++ source/extensions/Sekai.Interop.Desktop/Sekai.Interop.Desktop.csproj
@@ -4,6 +4,11 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Sekai.Interop.Common\Sekai.Interop.Common.csproj" />
+  </ItemGroup>
 
 </Project>
